finish service for buying using cash

+++ ApplicationCore/AssetAggregate/BankSavingAssetAggregate/DTOs/CreateNewBankSavingAssetDto.cs
@@ -7,7 +7,7 @@ namespace ApplicationCore.AssetAggregate.BankSavingAssetAggregate.DTOs
         public string Name { get; set; }
         public string BankCode { get; set; }
         public DateTime InputDay { get; set; }
-        public double InputMoneyAmount { get; set; }
+        public decimal InputMoneyAmount { get; set; }
         public string InputCurrency { get; set; } = "VND";
         public bool IsGoingToReinState { get; set; }
         public string Description { get; set; }
